+++ deltafi-ui/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Changed
+
+- Reworked plugins page to make editing variables safer and more user-friendly (#203)
+
 ## [0.96.4]
 
 ### Changed